fixed overflow for large normalmaps integrations

+++ src/bni_normal_integration.h
@@ -26,8 +26,8 @@ std::vector<float> bni_pyramid(std::function<bool(QString s, int n)> progressed,
 								  int max_solver_iterations = 5000,
 								  int scale = 0);
 
-bool savePly(const QString &filename, int w, int h, std::vector<float> &z);
-bool saveTiff(const QString &filename, int w, int h, std::vector<float> &z);
-bool saveDepthMap(const QString &filename, int w, int h, std::vector<float> &z);
+bool savePly(const QString &filename, size_t w, size_t h, std::vector<float> &z);
+bool saveTiff(const QString &filename, size_t w, size_t h, std::vector<float> &z);
+bool saveDepthMap(const QString &filename, size_t w, size_t h, std::vector<float> &z);
 
 #endif // BNI_NORMAL_INTEGRATION_H
